lda/requirements.txt
gensim
spacy
NLP-python




nltk.download('stopwords')
python3 -m spacy download en